refactor(api): move reconcileSupOrganizationLearner route into prescription
