--- api/lib/application/sup-organization-learners/index.js
@@ -1,46 +1,0 @@
-import JoiDate from '@joi/date';
-import BaseJoi from 'joi';
-const Joi = BaseJoi.extend(JoiDate);
-
-import { sendJsonApiError, UnprocessableEntityError } from '../../../src/shared/application/http-errors.js';
-import { supOrganizationLearnerController } from './sup-organization-learner-controller.js';
-
-const register = async function (server) {
-  server.route([
-    {
-      method: 'POST',
-      path: '/api/sup-organization-learners/association',
-      config: {
-        handler: supOrganizationLearnerController.reconcileSupOrganizationLearner,
-        validate: {
-          options: {
-            allowUnknown: false,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'student-number': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                birthdate: Joi.date().format('YYYY-MM-DD').required(),
-                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-              },
-              type: 'sup-organization-learners',
-            },
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
-          },
-        },
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Elle réconcilie l’utilisateur à l’inscription d’un étudiant dans cette organisation',
-        ],
-        tags: ['api', 'sup-organization-learner'],
-      },
-    },
-  ]);
-};
-
-const name = 'sup-organization-learners-api';
-export { name, register };
