+++ api/src/prescription/learner-management/application/sup-organization-management-controller.js
@@ -108,11 +108,31 @@ const updateStudentNumber = async function (request, h) {
   return h.response().code(204);
 };
 
+const reconcileSupOrganizationLearner = async function (request, h) {
+  const userId = request.auth.credentials.userId;
+  const payload = request.payload.data.attributes;
+
+  const campaignCode = payload['campaign-code'];
+
+  const reconciliationInfo = {
+    userId,
+    studentNumber: payload['student-number'],
+    firstName: payload['first-name'],
+    lastName: payload['last-name'],
+    birthdate: payload['birthdate'],
+  };
+
+  await usecases.reconcileSupOrganizationLearner({ campaignCode, reconciliationInfo });
+
+  return h.response(null).code(204);
+};
+
 const supOrganizationManagementController = {
   getOrganizationLearnersCsvTemplate,
   importSupOrganizationLearners,
   replaceSupOrganizationLearners,
   updateStudentNumber,
+  reconcileSupOrganizationLearner,
 };
 
 export { supOrganizationManagementController };
