+++ api/src/prescription/learner-management/application/sup-organization-management-route.js
@@ -20,6 +20,38 @@ const register = async function (server) {
   server.route([
     {
       method: 'POST',
+      path: '/api/sup-organization-learners/association',
+      config: {
+        handler: supOrganizationManagementController.reconcileSupOrganizationLearner,
+        validate: {
+          options: {
+            allowUnknown: false,
+          },
+          payload: Joi.object({
+            data: {
+              attributes: {
+                'student-number': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                birthdate: Joi.date().format('YYYY-MM-DD').required(),
+                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              },
+              type: 'sup-organization-learners',
+            },
+          }),
+          failAction: (request, h) => {
+            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
+          },
+        },
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            '- Elle réconcilie l’utilisateur à l’inscription d’un étudiant dans cette organisation',
+        ],
+        tags: ['api', 'sup-organization-learner'],
+      },
+    },
+    {
+      method: 'POST',
       path: '/api/organizations/{organizationId}/sup-organization-learners/replace-csv',
       config: {
         pre: [
